+++ app/city/[cityId]/category/[categoryId]/firm/[firmId]/add_review/interfaces.ts
@@ -1,0 +1,7 @@
+import { AddReviewFields } from './constants';
+
+export interface AddReviewValues {
+  [AddReviewFields.Author]: string;
+  [AddReviewFields.Text]: string;
+  [AddReviewFields.Rating]: string;
+}
