+++ app/city/[cityId]/category/[categoryId]/firm/[firmId]/page.tsx
@@ -29,7 +29,7 @@ export interface FirmPageParams {
   };
 }
 
-export const HeroBackground = {
+const HeroBackground = {
   '3ebc7206-6fed-4ea7-a000-27a74e867c9a':
     'https://images.unsplash.com/photo-1550966871-3ed3cdb5ed0c?q=80&w=5070&auto=format&fit=crop&ixlib=rb-4.0.3&ixid=M3wxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8fA%3D%3D',
   '565ad1cb-b891-4185-ac75-24ab3898cf22':
